Initiate concrete execution if a wrapper method is missing (#392)
If a JVM class is overridden but a method is missing from the wrapper,
the engine will discard the path and fall back to concrete execution
instead of analysing the real JVM code graph.

This approach fixes the problem with methods that have been introduced
in newer JDKs. Now wrappers are mostly limited to Java 1.8 interfaces
and fail to analyze methods like `String::isBlank` or `String::lines`
when the code runs under JDK 11. Building graphs from the real JDK code
fails because the wrapper does not have private fields that the original
code uses.

TODO: to allow symbolic analysis of the code, missing methods should be
actually implemented in corresponding wrappers.