Allow variables to have as general type as possible.
For example, let's say we are testing method 'foo(List l)'.
We may want to pass an ArrayList to this method as an argument. To do that code generation first creates a variable that stores this ArrayList. However, we don't have to use ArrayList as the type of this variable. It is better to use List in this case.

This commit makes code generation use actual types of 'this' and parameters for such variables.

This change also makes it possible to easily work with anonymous classes: we create an instance of an anonymous class, but we store it in a variable of its supertype. This way we are able to work with this anonymous class instance through the interface of its superclass. Thus, no reflection is needed for anonymous class variables and the code looks cleaner.